+++ ov/profiles.go
@@ -424,8 +424,8 @@ func (c *OVClient) SubmitNewProfile(p ServerProfile) (err error) {
 	}
 
 	// power off the server so that we can add to SP
-	if server.Name != "" {
-		server.PowerOff()
+	if server.Name != "" && server.PowerState == "on" {
+		return errors.New("Server Hardware must be powered off to assign to the server profile")
 	}
 
 	var emptyMgmtProcessorsStruct ManagementProcessors